Voting + credentials moved
I have added post voting on the poststream page, so you can now vote
posts up and down, and this is listed on the post. It should then be
possible to sort posts by votes if we wanted too. Currently, there is
nothing stopping you from making as near infinite votes as the database
would allow on a post.

I am planning on using a cookie and logging posts people have voted on
in a session, but that may be easier once we get the Facebook login
sorted.