Modularised!
All pages now called through a single index.php which deals with the
header and footer. Split components up a bit to allow for easier
management of the interface without messing up the back-end stuff.

KNOWN ISSUE: Voting doesn't work as the $_GET variables don't seem to
pass through the index.php - maybe split the string with the location in
at the '?' and then add that on again afterward?